+++ cpp/client/OAIAnalyzeRecipeInstructions_200_response_ingredients_inner.cpp
@@ -75,10 +75,10 @@ QJsonObject OAIAnalyzeRecipeInstructions_200_response_ingredients_inner::asJsonO
     return obj;
 }
 
-double OAIAnalyzeRecipeInstructions_200_response_ingredients_inner::getId() const {
+qint32 OAIAnalyzeRecipeInstructions_200_response_ingredients_inner::getId() const {
     return m_id;
 }
-void OAIAnalyzeRecipeInstructions_200_response_ingredients_inner::setId(const double &id) {
+void OAIAnalyzeRecipeInstructions_200_response_ingredients_inner::setId(const qint32 &id) {
     m_id = id;
     m_id_isSet = true;
 }
