+++ ruby/lib/openapi_client/models/analyze_recipe_instructions200_response_ingredients_inner.rb
@@ -35,7 +35,7 @@ module OpenapiClient
     # Attribute type mapping.
     def self.openapi_types
       {
-        :'id' => :'Float',
+        :'id' => :'Integer',
         :'name' => :'String'
       }
     end
