+++ ruby/lib/openapi_client/models/analyze_recipe_instructions200_response_parsed_instructions_inner_steps_inner_ingredients_inner.rb
@@ -41,7 +41,7 @@ module OpenapiClient
     # Attribute type mapping.
     def self.openapi_types
       {
-        :'id' => :'Float',
+        :'id' => :'Integer',
         :'name' => :'String',
         :'localized_name' => :'String',
         :'image' => :'String'
